stream-processing/producer/requirements.txt
kafka-python
hdfs
pandas